requirements.txt
# Core dependencies
langchain>=0.0.340
langchain-google-genai>=0.0.6
sentence-transformers>=2.2.2
chromadb>=0.4.18
requests>=2.31.0
gradio>=4.4.0
python-dotenv>=1.0.0
tqdm>=4.66.1
pydantic>=2.0,<3.0

# Vector store and embeddings
numpy>=1.24.3
scikit-learn>=1.3.0
torch>=2.0.1
transformers>=4.33.2

# Required for Hugging Face Spaces
fastapi<0.99.0
httpx>=0.24.1
python-multipart>=0.0.6
uvicorn>=0.21.0
aiofiles>=22.0
websockets>=11.0


